Refactor Zuora error handling to separate module
Moved error generation logic from zuoraClient.ts to a new zuoraErrorHandler.ts module. This improves code organization and separation of concerns by isolating error handling functionality.
